convert_model.py
#!/usr/bin/env python3
"""
Convert old incompatible model to new compatible format.
This creates a dummy BUT FUNCTIONAL model with the same interface.
WARNING: This model is not trained on real data - it's a placeholder!
"""

import joblib
import os
from sklearn.ensemble import RandomForestClassifier
from sklearn.tree import DecisionTreeClassifier
import numpy as np

def create_compatible_model():
    """
    Create a new RandomForest model compatible with current scikit-learn.
    This uses dummy training data with the expected feature structure.
    Replace this with real training data when available!
    """
    print("Creating a compatible RandomForest model...")
    
    # Create dummy training data matching the app's expected features
    # Features: [gender_binary, hemoglobin, mch, mchc, mcv]
   # Realistic training data for anemia detection
    X_train = np.array([
        # Anemic samples (label=1)
        [0, 10.0, 25.0, 30.0, 70.0],  # Female, low hemoglobin
        [1, 11.0, 24.0, 31.0, 72.0],  # Male, low hemoglobin
        [0, 9.5, 23.0, 29.0, 68.0],   # Female, very low
        [1, 10.5, 26.0, 30.0, 75.0],  # Male, low
        [0, 11.0, 25.0, 30.5, 71.0],  # Female, borderline
        [1, 10.0, 24.5, 30.0, 74.0],  # Male, low
        # Non-anemic samples (label=0)
        [0, 13.0, 28.0, 33.0, 85.0],  # Female, normal
        [1, 14.0, 29.0, 34.0, 88.0],  # Male, normal
        [0, 12.5, 27.5, 32.5, 84.0],  # Female, normal
        [1, 13.5, 28.5, 33.5, 87.0],  # Male, normal
        [0, 14.5, 30.0, 35.0, 90.0],  # Female, high
        [1, 15.0, 31.0, 36.0, 92.0],  # Male, high
    ])
    
    # Labels: 1 = Anemic, 0 = Not Anemic
    y_train = np.array([1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0])
    
    # Create and train RandomForest model
    model = RandomForestClassifier(
        n_estimators=10,
        max_depth=5,
        random_state=42,
        n_jobs=-1
    )
    
    model.fit(X_train, y_train)
    
    print(f"Model trained with accuracy: {model.score(X_train, y_train):.2f}")
    
    return model

def convert_old_model():
    """
    Force create a new compatible model.
    """
    model_path = 'model/random_forest_classifier.pkl'
    backup_path = 'model/random_forest_classifier.pkl.backup'
    
    # Backup the old model if exists
    if os.path.exists(model_path):
        import shutil
        if not os.path.exists(backup_path):
            shutil.copy(model_path, backup_path)
            print(f"Backed up old model to {backup_path}")
        else:
            print("Backup already exists, skipping backup")
    
    # Always create a new compatible model
    print("Creating a new compatible model...")
    new_model = create_compatible_model()
    
    # Save the new model
    os.makedirs('model', exist_ok=True)
    joblib.dump(new_model, model_path)
    print(f"✓ New model saved to {model_path}")
    
    return new_model

if __name__ == '__main__':
    convert_old_model()
    print("\nModel conversion complete!")
    print("Testing the model...")
    
    model = joblib.load('model/random_forest_classifier.pkl')
    
    # Test prediction
    test_data = np.array([[0, 12.0, 27.0, 32.0, 83.0]])  # Healthy female
    prediction = model.predict(test_data)
    print(f"Test prediction for healthy female: {prediction[0]} (0=Not Anemic, 1=Anemic)")
    
    test_data2 = np.array([[0, 10.0, 25.0, 30.0, 70.0]])  # Anemic female
    prediction2 = model.predict(test_data2)
    print(f"Test prediction for anemic female: {prediction2[0]} (0=Not Anemic, 1=Anemic)")
